cmd/search/pathindex: Pull pathIndex into its own file
So it's more obviously independent of executeGrep and its Index type.
